Separated rotational cipher generation and added comments
Separated the code that generates the map for the rotational cipher into its own method.
Added a private variable in the RotationalCipher class that holds the map for the pre-rotated and rotated characters.
This variable is populated when the constructor is called.
Also added some comments to the top of the custom methods.